#145: port _output and _dependency methods to adjoint.py
